Rework widget designs to use modern aesthetic
