+++ Eurofurence.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -33,8 +33,8 @@
         "repositoryURL": "https://github.com/firebase/firebase-ios-sdk.git",
         "state": {
           "branch": null,
-          "revision": "ccb2bccb82e62ad9e8a4e757a43ce6a04ceae1c9",
-          "version": "8.14.0"
+          "revision": "78f7087fd5d48eb7c36e299f330b6dddccd647b2",
+          "version": "8.12.1"
         }
       },
       {
@@ -42,8 +42,8 @@
         "repositoryURL": "https://github.com/google/GoogleAppMeasurement.git",
         "state": {
           "branch": null,
-          "revision": "9b74e6a3f875208c04c22b39b580a21d4de0f4f9",
-          "version": "8.14.0"
+          "revision": "6cc2991c11872510a5314bc112cc7558dd9d046a",
+          "version": "8.12.0"
         }
       },
       {
@@ -78,8 +78,8 @@
         "repositoryURL": "https://github.com/google/gtm-session-fetcher.git",
         "state": {
           "branch": null,
-          "revision": "eca9404a18f53727e4698211aaf2615eb93b962a",
-          "version": "1.7.1"
+          "revision": "bc6a19702ac76ac4e488b68148710eb815f9bc56",
+          "version": "1.7.0"
         }
       },
       {
@@ -105,8 +105,8 @@
         "repositoryURL": "https://github.com/ShezHsky/ObservedObject.git",
         "state": {
           "branch": null,
-          "revision": "176df11a5bcb4e5bc04d55c4d24db800b4e6cf69",
-          "version": "0.1.0"
+          "revision": "3e7cb70ac5a8d28430638d81d374bb6fe7925f2c",
+          "version": "0.2.0"
         }
       },
       {
