+++ EventsWidget/Resources/Color+Catalogue.swift
@@ -2,8 +2,10 @@ import SwiftUI
 
 extension Color {
     
-    static var widgetBackground: Color {
-        Color("WidgetBackground")
-    }
+    static let widgetContentForegroundPrimary = Color("WidgetContentForegroundPrimary")
+    static let widgetContentForegroundSecondary = Color("WidgetContentForegroundSecondary")
+    
+    static let widgetContentBackgroundGradientStart = Color("WidgetContentBackgroundGradientStart")
+    static let widgetContentBackgroundGradientEnd = Color("WidgetContentBackgroundGradientEnd")
     
 }
