--- EventsWidget/Views/Event/EventEndTime.swift
@@ -1,18 +1,0 @@
-import SwiftUI
-
-struct EventEndTime: View {
-    
-    var formattedEndTime: String
-    
-    init(_ formattedEndTime: String) {
-        self.formattedEndTime = formattedEndTime
-    }
-    
-    var body: some View {
-        Text(formattedEndTime)
-            .lineLimit(1)
-            .font(.caption2)
-            .foregroundColor(.secondary)
-    }
-    
-}
