+++ EventsWidget/Views/EventsWidgetEntryView.swift
@@ -1,0 +1,124 @@
+import EurofurenceKit
+import EventsWidgetLogic
+import SwiftUI
+import WidgetKit
+
+struct EventsWidgetEntryView: View {
+    
+    @Environment(\.widgetFamily) private var family: WidgetFamily
+    
+    var entry: EventTimelineEntry
+    
+    var body: some View {
+        switch family {
+        case .systemSmall:
+            SmallEventsWidget(entry: entry)
+                .springboardStyle()
+            
+        case .systemMedium:
+            MediumEventsWidget(entry: entry)
+                .springboardStyle()
+            
+        default:
+            fatalError("Unsupported widget family (\(family))")
+        }
+    }
+    
+}
+
+struct EventsWidget_Previews: PreviewProvider {
+    
+    static var previews: some View {
+        let events: [EventViewModel] = [
+            EventViewModel(
+                id: UUID().uuidString,
+                title: "Trans Meet-Up",
+                location: "Nizza",
+                formattedStartTime: "13:00",
+                formattedEndTime: "14:30"
+            ),
+            
+            EventViewModel(
+                id: UUID().uuidString,
+                title: "Dealer's Den",
+                location: "Dealer's Den - Convention Center Foyer 3",
+                formattedStartTime: "13:30",
+                formattedEndTime: "15:00"
+            ),
+            
+            EventViewModel(
+                id: UUID().uuidString,
+                title: "Funny Animals and Amerimanga in Sonic the Hedgehog Archie Series",
+                location: "Nizza",
+                formattedStartTime: "17:30",
+                formattedEndTime: "18:30"
+            ),
+            
+            EventViewModel(
+                id: UUID().uuidString,
+                title: "Fursuit Photoshoot Registration",
+                location: "Fursuit Photoshoot Registration - Estrel Hall B",
+                formattedStartTime: "19:00",
+                formattedEndTime: "20:30"
+            ),
+            
+            EventViewModel(
+                id: UUID().uuidString,
+                title: "International Snack Exchange",
+                location: "ECC Room 3",
+                formattedStartTime: "22:00",
+                formattedEndTime: "00:30"
+            )
+        ]
+        
+        let smallEntry = EventTimelineEntry(
+            date: Date(),
+            content: .events(viewModels: Array(events.prefix(2)), additionalEventsCount: 7),
+            context: EventTimelineEntry.Context(category: .upcoming, isFavouritesOnly: false)
+        )
+        
+        let mediumEntry = EventTimelineEntry(
+            date: Date(),
+            content: .events(viewModels: Array(events.prefix(3)), additionalEventsCount: 6),
+            context: EventTimelineEntry.Context(category: .upcoming, isFavouritesOnly: false)
+        )
+        
+        let noEvents = EventTimelineEntry(
+            date: Date(),
+            content: .events(viewModels: [], additionalEventsCount: 0),
+            context: EventTimelineEntry.Context(category: .upcoming, isFavouritesOnly: false)
+        )
+        
+        Group {
+            EventsWidgetEntryView(entry: smallEntry)
+                .previewContext(WidgetPreviewContext(family: .systemSmall))
+            
+            EventsWidgetEntryView(entry: noEvents)
+                .previewContext(WidgetPreviewContext(family: .systemSmall))
+            
+            EventsWidgetEntryView(entry: mediumEntry)
+                .previewContext(WidgetPreviewContext(family: .systemMedium))
+            
+            EventsWidgetEntryView(entry: noEvents)
+                .previewContext(WidgetPreviewContext(family: .systemMedium))
+        }
+        .previewDisplayName("Light")
+        .environment(\.colorScheme, .light)
+        
+        Group {
+            EventsWidgetEntryView(entry: smallEntry)
+                .previewContext(WidgetPreviewContext(family: .systemSmall))
+            
+            EventsWidgetEntryView(entry: noEvents)
+                .previewContext(WidgetPreviewContext(family: .systemSmall))
+            
+            EventsWidgetEntryView(entry: mediumEntry)
+                .previewContext(WidgetPreviewContext(family: .systemMedium))
+            
+            EventsWidgetEntryView(entry: noEvents)
+                .previewContext(WidgetPreviewContext(family: .systemMedium))
+        }
+        .environment(\.colorScheme, .dark)
+    }
+    
+}
