--- EventsWidget/Views/Widget/CategoryPlaceholderText.swift
@@ -1,26 +1,0 @@
-import EventsWidgetLogic
-import SwiftUI
-
-struct CategoryPlaceholderText: View {
-    
-    var category: EventCategory
-    
-    var body: some View {
-        text
-            .font(.caption)
-            .multilineTextAlignment(.center)
-            .foregroundColor(.primary)
-    }
-    
-    @ViewBuilder
-    private var text: some View {
-        switch category {
-        case .upcoming:
-            Text("No upcoming events")
-            
-        case .running:
-            Text("No running events")
-        }
-    }
-    
-}
